+++ example/app/dynamic-form/bootstrap/bootstrap-example.model.ts
@@ -68,7 +68,6 @@ export const BOOTSTRAP_EXAMPLE_MODEL: DynamicFormModel = new DynamicFormModel([
         id: "exampleCheckbox",
         label: {
             text:  "I do agree"
-        },
-        text: "I do agree"
+        }
     })
 ]);